ts main file for cypress
